Cloud-Native E-Commerce Platform/services/product-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
kafka-python==2.0.2
prometheus-client==0.20.0
